Give up on sad path handling as it's taking too long

+++ app/Http/Controllers/API/Auth/LoginController.php
@@ -46,8 +46,12 @@ class LoginController extends Controller
    */
   public function logout(Request $request): JsonResponse
   {
-    $request->user()->currentAccessToken()->delete();
+    if ($request->user()) {
+      $request->user()->currentAccessToken()->delete();
 
-    return response()->json(['message' => 'Logged out successfully'], 200);
+      return response()->json(['message' => 'Logged out successfully'], 200);
+    }
+
+    return response()->json(['error' => 'User is not logged in'], 401);
   }
 }
